ui: Reorg slightly and add plumbing for a faster route to list steps
